working clustering in. version bump to 1.0.0!
