+++ rfpipe/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.9.9'
+__version__ = '1.0.0'
