+++ setup.py
@@ -19,7 +19,8 @@ AUTHOR = 'Casey Law'
 
 # What packages are required for this module to be executed?
 REQUIRED = ['numpy', 'scipy', 'pwkit', 'sdmpy', 'pyfftw', 'bokeh',
-            'attrs', 'future', 'astropy<3.0', 'pyyaml', 'numba', 'HDBSCAN'],
+            'attrs', 'future', 'astropy<3.0', 'pyyaml', 'numba', 'hdbscan',
+            'seaborn'],
             # optional 'rfgpu', 'distributed'
 
 # The rest you shouldn't have to touch too much :)
